added a TableConverter to assist transferring tables between different version.
This converter consists of a CSV rule file and a python script:
1) the CSV file describes all the changes in a consistent manner from one version to another and can be customised to satisfy specific needs. The rules are quite self-explanatory: basically, it describes, when from a version to another, what action will be conducted for which file, which variable and which position.
2) the python script includes all functions to carry out the conversion and the basic inputs can be provided in its ending part.

The usage is simple: define the rules, put tables in two folders (‘from’ and ’to’), and run the script.

At this stage, the converter only supports three actions (which I find enough for now):
1) renaming, 2) deletion and 3) addition.
More actions may be added depending on our development needs.